Made animation cooler!
The typing animation previously typed once then stopped. Now it types and untypes a new word every 15 seconds or so. Array of words passed in from Sanity.

Still some issues, mainly the subheading times out very occasionally and nothing is shown, I'm not sure why this is. Also "setInterval" is not great as I learned JS timers don't run on a seperate thread so if the thread is busy  this might cause the timer to become inaccurate(!). So used date instead.

Another issue is the JS timer and CSS timer are seperate so I have to change two timers to change the whole animation duration which isn't good. Can fix this by putting animation in JSS instead of SCSS and pass timer through as argument