Trying to fix 'VerticalPaddedContainer' bug
Vs code file show one folder called 'Wrappers' in src with 'verticalPaddedContainer' file and 'externalLink' file. Github show two folders 'Wrappers' and 'wrappers' each with one of the files. This is causing Netlify to fail when building as it cant find the folder from imports.

Recreated the file with different name and put it in the 'Wrappers' folder